fix: provide default watch host
